+++ client-app/app/controllers/index.js
@@ -1,9 +1,20 @@
 import Controller from "@ember/controller";
 import { ajax } from "client-app/lib/utilities";
 import { observer, computed } from "@ember/object";
+import Preload from "client-app/lib/preload";
 
 export default Controller.extend({
+  showDebug: true,
+  showInfo: true,
+  showWarn: true,
+  showErr: true,
+  showFatal: true,
+  search: "",
   currentMessage: Em.computed.alias("model.currentMessage"),
+
+  showSettings: computed(function() {
+    return Preload.get("patterns_enabled");
+  }),
 
   resizePanels(amount) {
     Em.$("#bottom-panel").css("height", amount - 13);
@@ -69,7 +80,7 @@ export default Controller.extend({
     }
   },
 
-  filterChanged: observer(
+  filter: computed(
     "showDebug",
     "showInfo",
     "showWarn",
@@ -85,13 +96,18 @@ export default Controller.extend({
 
       // always show unknown, rare
       filter.push(5);
-      const model = this.get("model");
-      model.set("filter", filter);
-      if (this.get("initialized")) {
-        model.reload().then(() => this.updateSelectedMessage());
-      }
+      return filter;
     }
   ),
+
+  filterChanged: observer("filter.length", function() {
+    const filter = this.get("filter");
+    const model = this.get("model");
+    model.set("filter", filter);
+    if (filter && this.get("initialized")) {
+      model.reload().then(() => this.updateSelectedMessage());
+    }
+  }),
 
   searchChanged: observer("search", function() {
     const search = this.get("search");