+++ client-app/app/helpers/or.js
@@ -1,0 +1,7 @@
+import { helper } from "@ember/component/helper";
+
+export function or(params) {
+  return params.some(p => p);
+}
+
+export default helper(or);